Add files via upload
New signature MUTECOMPILER is defined with all the global variables and functions that are part of Structure Mutecompiler,
New structure Mutecompiler has two core functions silencecompiler and unsilencecompier.
a.	silencecompiler function mutes the compiler messages by saving the current printing limits in a ref cell and then set them all to zero.
b.	unsilencecompiler function unmutes the compiler messages by restoring the printing limits.
c.	dummyfn function which does nothing is created and invoked to preload the Mutecompiler structure before the script is passed to evalloop, this is to supress the structure auto-loading logs in the script results.