Add files via upload
A new function (useScriptFile) is added to Backend.Interact structure, which takes the file name and its content as a stream and process the stream by passing it to EvalLoop.evalStream. The compiler messages are muted and unmuted before the processing of the file.

a)	New function declaration is added to interact.sig,
b)	New function definition is added to interact.sml.